[not verified] process the connections list at effects level
