+++ projects/plugins/jetpack/extensions/plugins/publicize/hooks/use-social-media-connections/index.js
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { useDispatch, useSelect } from '@wordpress/data';
-import { store as editorStore } from '@wordpress/editor';
 
 /**
  * Hooks to deal with the social media connections.
@@ -11,9 +10,7 @@ import { store as editorStore } from '@wordpress/editor';
  */
 
 export default function useSocialMediaConnections() {
-	const { editPost } = useDispatch( editorStore );
 	const { refreshConnectionTestResults: refresh } = useDispatch( 'jetpack/publicize' );
-
 	const connections = useSelect( select => select( 'jetpack/publicize' ).getConnections(), [] );
 
 	return {
@@ -24,11 +21,7 @@ export default function useSocialMediaConnections() {
 				enabled: connection.id === id ? ! connection.enabled : connection.enabled,
 			} ) );
 
-			// Update post metadata.
-			editPost( { jetpack_publicize_connections } );
-
-			// Refresh jetpack/publicize store.
-			refresh();
+			refresh( jetpack_publicize_connections );
 		},
 
 		refresh,
