+++ projects/plugins/jetpack/extensions/plugins/publicize/store/actions.js
@@ -8,27 +8,35 @@ import { store as editorStore } from '@wordpress/editor';
  * Returns an action object used in signalling that
  * we're setting the Publicize connection test results.
  *
- * @param {Array} results - Connection test results.
+ * @param {Array} connections - Processed connections list.
  * @returns {object} Action object.
  */
-export function setConnectionTestResults( results ) {
+export function setConnectionTestResults( connections ) {
 	return {
 		type: 'SET_CONNECTION_TEST_RESULTS',
-		results,
+		connections,
 	};
 }
 
 /**
  * Returns an action object used in signalling that
- * we're refreshing the Publicize connection test results.
- * It picks the connections from the post meta.
+ * we're refreshing the Publicize connection.
  *
+ * It accepts an optional `connections` parameter, which
+ * can be used to refresh the connections straight away.
+ * Otherwise, it will pick the connections from the post metadata.
+ *
+ * @param {Array} connections - Processed connections list.
  * @returns {object} Action object.
  */
-export function refreshConnectionTestResults() {
+export function refreshConnectionTestResults( connections = null ) {
+	connections = connections
+		? connections
+		: select( editorStore ).getEditedPostAttribute( 'jetpack_publicize_connections' ) || [];
+
 	return {
 		type: 'REFRESH_CONNECTION_TEST_RESULTS',
-		results: select( editorStore ).getEditedPostAttribute( 'jetpack_publicize_connections' ) || [],
+		connections,
 	};
 }
